// src/payload-types.ts
export interface Volunteer {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  email: string;
  volunteering?: (number | VolunteerReason)[] | null;
  additionalInfo?: string | null;
  county: string;
  constituency: string;
  ward: string;
  pollingStation?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "volunteer-reasons".
 */
export interface VolunteerReason {
  id: number;
  reason: string;
  volunteers?: {
    docs?: (number | Volunteer)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}